requirements.txt
imageio-ffmpeg
pillow
scipy
numpy<2
tqdm
pyaml
torch==2.3
torchvision
safetensors
opencv-python
psutil
huggingface-hub